# src/main/java/DTOs/LoginReqDTO.java
package DTOs;

import lombok.Data;

@Data
public class LoginReqDTO {
    private String password;
    private String username;

    public String getPassword() {
        return password;
    }

    public void setPassword(String password) {
        this.password = password;
    }

    public String getUsername() {
        return username;
    }

    public void setUsername(String username) {
        this.username = username;
    }
}